+++ modules/canvaskit/npm_build/types/index.d.ts
@@ -1819,9 +1819,10 @@ export interface Canvas extends EmbindObject<"Canvas"> {
      * @param bounds
      * @param backdrop
      * @param flags
+     * @param backdropFilterTileMode
      */
     saveLayer(paint?: Paint, bounds?: InputRect | null, backdrop?: ImageFilter | null,
-              flags?: SaveLayerFlag): number;
+              flags?: SaveLayerFlag, backdropFilterTileMode?: TileMode): number;
 
     /**
      * Scales the current matrix by sx on the x-axis and sy on the y-axis.